# openapi.yaml
openapi: 3.1.0
info:
  title: LibraTech API
  version: 1.0.0
  description: LibraTech API — katalog buku, kategori, dan peminjaman perpustakaan. Base URL `/api/v1`, auth Bearer (Sanctum).
servers:
- url: https://libratech.lambada.my.id/api
...
security:
- http: []
paths:
...
  /v1/admin/books/{book}:
    put:
      operationId: adminBook.update
      tags:
      - Admin — Books
      parameters:
      - name: book
        in: path
        required: true
        description: The book ID
        schema:
          type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: '`BookResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                    - $ref: '#/components/schemas/BookResource'
                    - type: object
                      required:
                      - category
                required:
                - data
        '404':
          $ref: '#/components/responses/ModelNotFoundException'
        '401':
          $ref: '#/components/responses/AuthenticationException'
        '422':
          $ref: '#/components/responses/ValidationException'
        '403':
          $ref: '#/components/responses/AuthorizationException'
    delete:
      operationId: adminBook.destroy
      tags:
      - Admin — Books
      parameters:
      - name: book
        in: path
        required: true
        description: The book ID
        schema:
          type: integer
      responses:
        '204':
          description: No content
        '403':
          $ref: '#/components/responses/AuthorizationException'
        '404':
          $ref: '#/components/responses/ModelNotFoundException'
        '401':
          $ref: '#/components/responses/AuthenticationException'
...
  /v1/admin/loans/{loan}/approve:
    patch:
      operationId: loanAdmin.approve
      tags:
      - Admin — Loans
      parameters:
      - name: loan
        in: path
        required: true
        description: The loan ID
        schema:
          type: integer
      responses:
        '200':
          description: '`LoanResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                    - $ref: '#/components/schemas/LoanResource'
                    - type: object
                      required:
                      - book
                      - user
                required:
                - data
        '422':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                - message
        '403':
          $ref: '#/components/responses/AuthorizationException'
        '404':
          $ref: '#/components/responses/ModelNotFoundException'
        '401':
          $ref: '#/components/responses/AuthenticationException'
...
components:
  securitySchemes:
    http:
      type: http
      scheme: bearer
  schemas:
    BookResource:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        category:
          $ref: '#/components/schemas/CategoryResource'
        stock:
          type: integer
        is_available:
          type: boolean
        description:
          type:
          - string
          - 'null'
        cover_url:
          type:
          - string
          - 'null'
        published_year:
          type:
          - integer
          - 'null'
      required:
      - id
      - title
      - author
      - isbn
      - stock
      - is_available
      - description
      - cover_url
      - published_year
      title: BookResource
    CategoryResource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required:
      - id
      - name
      - slug
      title: CategoryResource
    LoanResource:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        book:
          allOf:
          - $ref: '#/components/schemas/BookResource'
          - type: object
            required:
            - category
        user:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
          required:
          - id
          - name
          - email
        requested_at:
          type:
          - string
          - 'null'
        approved_at:
          type:
          - string
          - 'null'
        due_at:
          type:
          - string
          - 'null'
        returned_at:
          type:
          - string
          - 'null'
        rejection_reason:
          type:
          - string
          - 'null'
      required:
      - id
      - status
      - requested_at
      - approved_at
      - due_at
      - returned_at
      - rejection_reason
      title: LoanResource
...
    UpdateBookRequest:
      type: object
      properties:
        category_id:
          type: integer
        title:
          type: string
          maxLength: 255
        author:
          type: string
          maxLength: 255
        isbn:
          type: string
          maxLength: 20
        stock:
          type: integer
          minimum: 0
        description:
          type:
          - string
          - 'null'
        cover:
          type:
          - string
          - 'null'
          format: binary
          contentMediaType: application/octet-stream
          description: 'Maximum file size: 2048 kilobytes.'
        published_year:
          type:
          - integer
          - 'null'
          minimum: 1000
          maximum: 2027
      required:
      - category_id
      - title
      - author
      - isbn
      - stock
      title: UpdateBookRequest
  responses:
    ValidationException:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Errors overview.
              errors:
                type: object
                description: A detailed description of each field that failed validation.
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
            - message
            - errors
    AuthenticationException:
      description: Unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error overview.
            required:
            - message
    ModelNotFoundException:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error overview.
            required:
            - message
    AuthorizationException:
      description: Authorization error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error overview.
            required:
            - message